+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.3] - 2026-08-12
+
+### Fixed
+
+- Prevent a fatal error when the "Virtual Media Folders" parent plugin is missing or older than 2.0.0; show an admin notice instead.
+
+### Changed
+
+- Tested up to WordPress 7.1.
+
 ## [0.3.2] - 2026-07-12
 
 ### Security